# openapi.yaml
openapi: 3.0.1
info:
  title: Paperless Rest Server
  version: '1.0'
paths:
...
  /api/search/autocomplete:
    get:
      tags:
        - Search
      operationId: AutoComplete
      parameters:
        - name: term
          in: query
          style: form
          schema:
            type: string
        - name: limit
          in: query
          style: form
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              examples:
                default:
                  value: '["term1", "term3", "term6", "term4"]'
...